better fuzzing

+++ bin/getflickr.php
@@ -5,18 +5,23 @@ $flickr = 'http://api.flickr.com/services/rest/?method=flickr.photos.search&api_
 // farm server-id id secret
 $photo_url = 'http://farm%s.staticflickr.com/%s/%s_%s.jpg';
 
-$process = dirname(__FILE__)."/test.php";
+$process = realpath(dirname(__FILE__)."/test.php");
+$kitty = realpath(dirname(__FILE__)."/../flickr");
 
 function getFlickr($terms, $page) {
-    global $flickr, $photo_url, $process;
+    global $flickr, $photo_url, $process, $kitty;
     $url = sprintf($flickr, $terms, $page);
     $raw = file_get_contents($url);
     $res = json_decode($raw);
     
     foreach ($res->photos->photo as $photo) {
         $url = sprintf($photo_url, $photo->farm, $photo->server, $photo->id, $photo->secret);
-        $local_name = "./flickr/{$photo->id}.jpg";
-        file_put_contents($local_name, file_get_contents($url));
+        $local_name = "$kitty/{$photo->id}.jpg";
+        
+        if (!file_exists($local_name)) {
+            file_put_contents($local_name, file_get_contents($url));
+        }
+
         print("php $process $local_name\n");
         system("php $process $local_name\n");
     }
